# monotickets-api/docs/contracts/OPENAPI_monotickets.yaml
openapi: 3.0.0
paths:
...
  /scanner/validate:
    post:
      operationId: ScannerController_validate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanValidationDto"
      responses:
        "200":
          description: QR validated successfully
          content:
            application/json:
              schema:
                example:
                  valid: true
                  status: VALID
                  guest:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    fullName: Juan Pérez
                    guestCount: 2
                    inviteType: STANDARD
                  scan:
                    id: 550e8400-e29b-41d4-a716-446655440001
                    scannedAt: 2025-12-05T15:00:00.000Z
                    scannedBy: 550e8400-e29b-41d4-a716-446655440002
                  message: Acceso permitido
        "400":
          description: Invalid QR or duplicate scan
      security: &a1
        - bearer: []
      summary: Validate QR code and register scan
      tags: &a2
        - scanner
  /scanner/confirm-partial:
    post:
      operationId: ScannerController_confirmPartial
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPartialEntryDto"
      responses:
        "200":
          description: Partial entry confirmed successfully
          content:
            application/json:
              schema:
                example:
                  success: true
                  scan:
                    id: 550e8400-e29b-41d4-a716-446655440001
                    enteredNames:
                      - Juan Pérez
                      - María García
                    remainingCount: 2
                    status: VALID_PARTIAL
                  message: Acceso permitido - Quedan 2 personas
        "400":
          description: Invalid request or too many names
        "404":
          description: Invitation not found
      security: *a1
      summary: Confirm partial entry for group invitation
      tags: *a2
  /scanner/history/{eventId}:
    get:
      operationId: ScannerController_getHistory
      parameters:
        - name: eventId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Scan history retrieved successfully
      security: *a1
      summary: Get scan history for an event
      tags: *a2
  /scanner/sync:
    post:
      operationId: ScannerController_sync
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncScansDto"
      responses:
        "200":
          description: Scans synced successfully
          content:
            application/json:
              schema:
                example:
                  synced: 45
                  failed: 5
                  results: []
      security: *a1
      summary: Sync offline scans
      tags: *a2
...
info:
  title: Monotickets Platinum API
  description: API oficial para Monotickets Platinum (invited events, RSVPs,
    delivery y scanner).
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ScanValidationDto:
      type: object
      properties:
        qrToken:
          type: string
          description: QR token to validate
          example: EV2025-001-GUEST-123-ABC
        eventId:
          type: string
          description: Event ID
          example: 550e8400-e29b-41d4-a716-446655440000
        scannedBy:
          type: string
          description: Staff user ID who is scanning
          example: 550e8400-e29b-41d4-a716-446655440001
        scannedAt:
          type: string
          description: Timestamp when scan occurred
          example: 2025-12-05T15:30:00.000Z
      required:
        - qrToken
        - eventId
        - scannedBy
        - scannedAt
    ConfirmPartialEntryDto:
      type: object
      properties:
        invitationId:
          type: string
          description: Invitation ID
          example: 123e4567-e89b-12d3-a456-426614174000
        enteredNames:
          description: Array of names of people who are entering
          example:
            - Juan Pérez
            - María García
          type: array
          items:
            type: string
        scannedBy:
          type: string
          description: ID of the staff member scanning
          example: 456e7890-e89b-12d3-a456-426614174111
      required:
        - invitationId
        - enteredNames
        - scannedBy
    OfflineScanDto:
      type: object
      properties:
        id:
          type: string
        qrToken:
          type: string
        eventId:
          type: string
        scannedBy:
          type: string
        scannedAt:
          type: string
      required:
        - id
        - qrToken
        - eventId
        - scannedBy
        - scannedAt
    SyncScansDto:
      type: object
      properties:
        scans:
          description: Array of offline scans to sync
          type: array
          items:
            $ref: "#/components/schemas/OfflineScanDto"
      required:
        - scans
...